refactor : Position 에게 위치에 관한 책임 위임

+++ src/main/java/racingcar/constant/Constant.java
@@ -5,7 +5,7 @@ public class Constant {
     public static final String DASH = "-";
     public static final int START_POSITION = 0;
     public static final int LEAST_MOVE_CONDITION = 4;
-    public static final int ZERO = 0;
+    public static final long ZERO = 0;
     public static final int BOUND = 10;
     public static final String ENTER = "\n";
 }
